fix: trade_paramsを別ファイルに分離
